VKT(Frontend): OPHVKTKEH-95 enrollment utils for partial exams related logic
