+++ frontend/packages/vkt/src/interfaces/common/enrollment.ts
@@ -1,0 +1,9 @@
+export interface PartialExamsAndSkills {
+  oralSkill: boolean;
+  textualSkill: boolean;
+  understandingSkill: boolean;
+  speakingPartialExam: boolean;
+  speechComprehensionPartialExam: boolean;
+  writingPartialExam: boolean;
+  readingComprehensionPartialExam: boolean;
+}
